requirements.txt
requests
coverage
flake8
mock
pyflakes
pytest
vcrpy
